requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mysql-connector-python==9.0.0
requests==2.32.3
scikit-learn==1.5.1
numpy==2.0.1
pandas==2.2.2
python-dateutil==2.9.0.post0
python-docx>=0.8.11
gunicorn>=21.2.0